// src/models/uchilishta-model/uchilishta.ts
export class Uchilishta {
    Address: string;
    Description: string;
    Email: string;
    FB: string;
    Fax: string;
    Founded: string;
    ID: string;
    Keywords: string;
    LocationX: string;
    LocationY: string;
    Name: string;
    SchoolType: string;
    Shortcut: string;
    TW: string;
    Telephone: string;
    TypeID: string;
    Website: string;
}